+++ demo-order-module/demo-order-base-domain/src/main/java/com/liyulin/demo/order/base/domain/mapper/OrderBillBaseMapper.java
@@ -1,11 +1,9 @@
 package com.liyulin.demo.order.base.domain.mapper;
 
-import java.math.BigInteger;
-
 import com.liyulin.demo.mybatis.mapper.ext.ExtMapper;
 import com.liyulin.demo.order.base.domain.entity.OrderBillEntity;
 import com.liyulin.demo.order.rpc.response.OrderBillRespBody;
 
-public interface OrderBillBaseMapper extends ExtMapper<OrderBillEntity, OrderBillRespBody, BigInteger> {
+public interface OrderBillBaseMapper extends ExtMapper<OrderBillEntity, OrderBillRespBody, Long> {
 
 }
