+++ demo-order-module/demo-order-base-domain/src/main/java/com/liyulin/demo/order/base/domain/mapper/OrderDeliveryInfoBaseMapper.java
@@ -1,12 +1,10 @@
 package com.liyulin.demo.order.base.domain.mapper;
 
-import java.math.BigInteger;
-
 import com.liyulin.demo.mybatis.mapper.ext.ExtMapper;
 import com.liyulin.demo.order.base.domain.entity.OrderDeliveryInfoEntity;
 import com.liyulin.demo.order.rpc.response.OrderDeliveryInfoRespBody;
 
 public interface OrderDeliveryInfoBaseMapper
-		extends ExtMapper<OrderDeliveryInfoEntity, OrderDeliveryInfoRespBody, BigInteger> {
+		extends ExtMapper<OrderDeliveryInfoEntity, OrderDeliveryInfoRespBody, Long> {
 
 }
